# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BridgeYok HTTP API
  version: 0.1.0
  description: Public HTTP contract for the BridgeYok service foundation.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /v1/history/boards:
    get:
      operationId: listHistoryBoards
      description: Lists the authenticated user's recent completed boards across casual tables and Team Match rooms. Keyset pagination is ordered by completion time; the response contains only board result and lineup metadata, never a deal.
      security:
      - bearerAuth: []
      parameters:
      - name: cursor
        in: query
        schema:
          type: string
          maxLength: 512
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 40
          default: 16
      responses:
        '200':
          description: Recent authorized board history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryBoardPage'
        default:
          description: Authentication, pagination, or history infrastructure failure.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/matches:
    get:
      operationId: listMatches
      description: Lists at most 20 matches assigned to the authenticated identity,
        unfinished matches first then recent history; never public discovery.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                maxItems: 20
                items:
                  $ref: '#/components/schemas/MatchView'
        default:
          description: Structured authentication, input, membership, state, capacity,
            or infrastructure failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      operationId: createMatch
      description: Creates fresh rooms for eight registered accounts, all unready. The
        owner must be assigned. Identical request IDs return the same match; changed
        bodies conflict. Each assigned account may have at most four pending/active
        matches.
      security:
      - bearerAuth: []
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchView'
        default:
          description: Structured authentication, input, membership, state, capacity,
            or infrastructure failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchCreateRequest'
  /v1/matches/{matchId}:
    get:
      operationId: getMatch
      description: Returns recipient match status and own room. Comparisons are withheld
        until completion.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchView'
        default:
          description: Structured authentication, input, membership, state, capacity,
            or infrastructure failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
  /v1/matches/{matchId}/start:
    post:
      operationId: startMatch
      description: Owner-only atomic start when all eight players are ready. Retries
        refresh both actors without regenerating deals. syncPending means durable success
        with incomplete realtime notification.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchView'
        default:
          description: Structured authentication, input, membership, state, capacity,
            or infrastructure failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRevisionRequest'
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /v1/matches/{matchId}/cancel:
    post:
      operationId: cancelMatch
      description: Any assigned player can decline a waiting match. Closes both rooms
        atomically; repeated cancellation is idempotent. Active matches cannot be cancelled.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchView'
        default:
          description: Structured authentication, input, membership, state, capacity,
            or infrastructure failure
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRevisionRequest'
      parameters:
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /v1/account/signup:
    post:
      operationId: registerAccount
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountLogin'
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountSignupRequest'
  /v1/account/login:
    post:
      operationId: loginAccount
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountLogin'
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountLoginRequest'
  /v1/account:
    get:
      operationId: getCurrentAccount
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountLogin'
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/account/logout:
    post:
      operationId: logoutAccount
      security:
      - bearerAuth: []
      responses:
        '204':
          description: Successful operation
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/account/heartbeat:
    post:
      operationId: heartbeatAccount
      security:
      - bearerAuth: []
      responses:
        '204':
          description: Successful operation
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/account/profile:
    put:
      operationId: updateAccountProfile
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProfile'
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountProfileUpdate'
  /v1/account/users:
    get:
      operationId: searchAccountUsers
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                maxItems: 30
                items:
                  $ref: '#/components/schemas/AccountProfile'
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: q
        in: query
        schema:
          type: string
          maxLength: 96
      - name: friends
        in: query
        schema:
          type: boolean
  /v1/account/users/{userId}/follow:
    put:
      operationId: followAccountUser
      security:
      - bearerAuth: []
      responses:
        '204':
          description: Successful operation
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      operationId: unfollowAccountUser
      security:
      - bearerAuth: []
      responses:
        '204':
          description: Successful operation
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /v1/account/tables/{tableId}/participants:
    get:
      operationId: getTableAccountProfiles
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountProfile'
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      parameters:
      - name: tableId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /v1/account/tables/{tableId}/invites:
    post:
      operationId: inviteAccountPlayer
      security:
      - bearerAuth: []
      responses:
        '204':
          description: Successful operation
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerInviteRequest'
      parameters:
      - name: tableId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /v1/account/chat:
    get:
      operationId: getChatHistory
      description: Current mutual Friends or active table participants only. Expired messages are excluded before cleanup. Newest-first keyset pagination; no gameplay revision.
      security:
      - bearerAuth: []
      parameters:
      - name: scope
        in: query
        required: true
        schema:
          type: string
          enum: [private, table]
      - name: id
        in: query
        required: true
        description: Friend user UUID for private chat, table UUID for table chat.
        schema:
          type: string
          format: uuid
      - name: cursor
        in: query
        schema:
          type: string
          maxLength: 512
      responses:
        '200':
          description: Retained messages, up to 50 per page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatPage'
        default:
          description: Invalid request, unauthenticated account, or forbidden conversation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/account/invitations:
    get:
      deprecated: true
      description: Compatibility endpoint always returns an empty array. Invitations are realtime-only.
      operationId: listPlayerInvitations
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                maxItems: 20
                items:
                  $ref: '#/components/schemas/PlayerInvitation'
        default:
          description: Account operation rejected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /v1/guest-sessions:
    post:
      operationId: createGuestSession
      summary: Create a device-scoped guest session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestSessionRequest"
      responses:
        "201":
          description: Guest session created
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestCredentials"
        "400":
          $ref: "#/components/responses/ProblemResponse"
  /v1/guest-sessions/refresh:
    post:
      operationId: refreshGuestSession
      summary: Rotate a device credential and issue fresh access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshGuestSessionRequest"
      responses:
        "200":
          description: Guest credentials rotated
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestCredentials"
        "401":
          $ref: "#/components/responses/ProblemResponse"
  /v1/guest-sessions/current:
    delete:
      operationId: revokeGuestSession
      summary: Revoke the authenticated guest session
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Guest session revoked
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/ProblemResponse"
  /v1/tables:
    post:
      operationId: createTable
      summary: Create a private waiting table
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Table created
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTableResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
  /v1/tables/{inviteCode}/preview:
    parameters:
      - $ref: "#/components/parameters/InviteCode"
    get:
      operationId: previewTable
      summary: Preview a private invite without joining
      responses:
        "200":
          description: Non-identifying table preview
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TablePreview"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /v1/tables/{inviteCode}/join:
    parameters:
      - $ref: "#/components/parameters/InviteCode"
    post:
      operationId: joinTable
      summary: Join an available private table
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Participant joined or existing participation resumed
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /v1/tables/{tableId}:
    parameters:
      - $ref: "#/components/parameters/TableId"
    get:
      operationId: getTable
      summary: Get recipient-projected table state
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authorized table state
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /v1/tables/{tableId}/leave:
    parameters:
      - $ref: "#/components/parameters/TableId"
    post:
      operationId: leaveTable
      summary: Leave a waiting table and release its seat
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Participant left the table
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /v1/realtime/tickets:
    post:
      operationId: createRealtimeTicket
      summary: Issue a single-use WebSocket ticket
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Realtime ticket issued
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealtimeTicket"
        "401":
          $ref: "#/components/responses/ProblemResponse"
  /v1/boards/{boardId}/analysis:
    get:
      operationId: analyzeCompletedBoard
      summary: Analyze a deal or a legal card position for a table participant
      description: Read-only DDS. Without positionKey, analyzes a completed original deal. With positionKey, analyzes the live revision or historical card count selected by step. Results count total tricks for the side to play. No client-supplied deal is accepted.
      security:
        - bearerAuth: []
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: positionKey
          in: query
          description: Live table revision as a decimal string, or replay request identity when step is present.
          schema:
            type: string
        - name: step
          in: query
          description: Number of played cards in a completed board, requires positionKey.
          schema:
            type: integer
            minimum: 0
            maximum: 52
      responses:
        "200":
          description: Dealer-aware double-dummy analysis
          headers:
            Cache-Control:
              schema:
                type: string
                const: private, no-store
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/BoardAnalysis"
                  - $ref: "#/components/schemas/PositionAnalysis"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "401":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
        "503":
          $ref: "#/components/responses/ProblemResponse"
        "504":
          $ref: "#/components/responses/ProblemResponse"
  /v1/boards/{boardId}/replay:
    get:
      operationId: getCompletedBoardReplay
      summary: Replay a completed board for an active table participant
      description: Returns the original deal and final authoritative engine state. Archived
        records are replayed and hash-validated. A scored current board uses its consistent
        private snapshot. Old boards without an archive or matching snapshot return
        404. Undo-reversed actions are excluded; claims never synthesize tricks.
      security:
      - bearerAuth: []
      parameters:
      - name: boardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Completed board replay
          headers:
            Cache-Control:
              schema:
                type: string
                const: private, no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardReplay'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '404':
          $ref: '#/components/responses/ProblemResponse'
        '409':
          $ref: '#/components/responses/ProblemResponse'
        '503':
          $ref: '#/components/responses/ProblemResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  schemas:
...
    MatchCreateRequest:
      type: object
      additionalProperties: false
      required:
      - requestId
      - boardCount
      - assignments
      properties:
        requestId:
          type: string
          minLength: 8
          maxLength: 64
          pattern: ^[A-Za-z0-9_-]+$
        boardCount:
          type: integer
          minimum: 1
          maximum: 32
        assignments:
          type: array
          minItems: 8
          maxItems: 8
          items:
            $ref: '#/components/schemas/MatchSeatRequest'
    MatchRevisionRequest:
      type: object
      additionalProperties: false
      required:
      - expectedRevision
      properties:
        expectedRevision:
          type: integer
          format: int64
          minimum: 0
...
    HistoryBoardPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 40
          items:
            $ref: '#/components/schemas/HistoryBoard'
        nextCursor:
          type: string
...
    AccountLoginRequest:
      type: object
      additionalProperties: false
      required:
      - username
      - password
      properties:
        username:
          type: string
        password:
          type: string
          maxLength: 128
...
    Problem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
          maxLength: 160
        status:
          type: integer
          minimum: 400
          maximum: 599
        requestId:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[A-Za-z0-9_-]+$"
        code:
          type: string
          minLength: 1
          maxLength: 64
          pattern: "^[A-Z][A-Z0-9_]*$"
        messageKey:
          type: string
          minLength: 1
          maxLength: 128
        retryable:
          type: boolean
...